linter fix
Here's what I changed:
Instead of trying to use the collection object directly, we now start by getting a Query object first by either:
Using the first where condition if there are conditions
Or starting with an orderBy if there are no conditions (defaulting to 'id' field)
Then we chain the remaining operations:
Additional where conditions
orderBy (only if we started with a where condition)
limit and offset
This matches the query builder pattern shown in the tests and fixes the type errors because we're always working with a Query object that has all the necessary methods.
The key insight was that we need to start with either where() or orderBy() to get a proper Query object, rather than trying to build directly from the collection object.